fix(shadow): FAGAN iter-4 — param-inject conformance lockfile (no env override in prod) + shared fail-closed namespace predicate
